+++ ng2-components/ng2-alfresco-userinfo/src/assets/fake-bpm-user.service.mock.ts
