+++ ng2-components/ng2-alfresco-userinfo/src/assets/fake-ecm-user.service.mock.ts
